pyproject.toml
[project]
name = "nightpaw"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "discord-py[voice]>=2.0",
    "python-dotenv",
    "aiosqlite",
    "aiohttp",
    "psutil>=7.2.2",
    "ephem>=4.2.1",
]
